services/mac: moved validation here
